Allow to use a custom cache instead of using the filesystem only
This commit adds the ability to specify a cache factory function instead of a folder
name only when creating an Authflow instance.

The cache factory function will take additional options and must return a cache. A cache
is an class instance that supports basic read and write operations.

The old behavior of the cache has been implemented by a FileCache. The FileCache will be
used as default and works like the old implementation used to.